small prof helper/arcdpsenums cleanup
