+++ GW2EIEvtcParser/EIData/Buffs/BuffInfoSolver.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using GW2EIEvtcParser.Interfaces;
 using GW2EIEvtcParser.ParsedData;
+using static GW2EIEvtcParser.ArcDPSEnums;
 using static GW2EIEvtcParser.ParserHelper;
 using static GW2EIEvtcParser.SkillIDs;
