+++ GW2EIEvtcParser/EIData/ProfHelpers/ProfHelper.cs
@@ -301,101 +301,9 @@ namespace GW2EIEvtcParser.EIData
             (int)ArcDPSEnums.MinionID.RuneMarkIGolem,
         };
 
-        internal static bool IsKnownMinionID(AgentItem minion, Spec spec)
+        internal static bool IsKnownMinionID(long id)
         {
-            if (minion.Type == AgentItem.AgentType.Gadget)
-            {
-                return false;
-            }
-            long id = minion.ID;
-            bool res = CommonMinions.Contains(id);
-            switch (spec)
-            {
-                //
-                case Spec.Elementalist:
-                case Spec.Tempest:
-                case Spec.Weaver:
-                case Spec.Catalyst:
-                    res |= ElementalistHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Necromancer:
-                case Spec.Scourge:
-                case Spec.Harbinger:
-                    res |= NecromancerHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Reaper:
-                    res |= NecromancerHelper.IsKnownMinionID(id);
-                    res |= ReaperHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Mesmer:
-                    res |= MesmerHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Chronomancer:
-                    res |= MesmerHelper.IsKnownMinionID(id);
-                    res |= ChronomancerHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Mirage:
-                    res |= MesmerHelper.IsKnownMinionID(id);
-                    res |= MirageHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Virtuoso:
-                    res |= MesmerHelper.IsKnownMinionID(id);
-                    res |= VirtuosoHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Thief:
-                    res |= ThiefHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Daredevil:
-                    res |= ThiefHelper.IsKnownMinionID(id);
-                    res |= DaredevilHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Deadeye:
-                    res |= ThiefHelper.IsKnownMinionID(id);
-                    res |= DeadeyeHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Specter:
-                    res |= ThiefHelper.IsKnownMinionID(id);
-                    res |= SpecterHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Engineer:
-                case Spec.Scrapper:
-                case Spec.Holosmith:
-                    res |= EngineerHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Mechanist:
-                    res |= EngineerHelper.IsKnownMinionID(id);
-                    res |= MechanistHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Ranger:
-                case Spec.Druid:
-                case Spec.Soulbeast:
-                case Spec.Untamed:
-                    res |= RangerHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Revenant:
-                case Spec.Herald:
-                case Spec.Vindicator:
-                    res |= RevenantHelper.IsKnownMinionID(id);
-                    break;
-                case Spec.Renegade:
-                    res |= RevenantHelper.IsKnownMinionID(id);
-                    res |= RenegadeHelper.IsKnownMinionID(id);
-                    break;
-                //
-                case Spec.Guardian:
-                case Spec.Dragonhunter:
-                case Spec.Firebrand:
-                case Spec.Willbender:
-                    res |= GuardianHelper.IsKnownMinionID(id);
-                    break;
-            }
-            return res;
+            return CommonMinions.Contains(id);
         }
 
         public static void ComputeMinionCombatReplayActors(AbstractSingleActor minion, AbstractSingleActor master, ParsedEvtcLog log, CombatReplay combatReplay)
