+++ GW2EIEvtcParser/ParserHelpers/ArcDPSEnums.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using GW2EIEvtcParser.EIData;
-using GW2EIEvtcParser.ParsedData;
 
 namespace GW2EIEvtcParser
 {
@@ -8,6 +6,64 @@ namespace GW2EIEvtcParser
     {
 
         public const int ArcDPSPollingRate = 300;
+
+        internal static class GW2Builds
+        {
+            internal const ulong StartOfLife = ulong.MinValue;
+            //
+            internal const ulong May2018Balance = 88541;
+            internal const ulong July2018Balance = 90455;
+            internal const ulong August2018Balance = 92069;
+            internal const ulong October2018Balance = 92715;
+            internal const ulong December2018Balance = 94051;
+            internal const ulong March2019Balance = 95535;
+            internal const ulong April2019Balance = 96406;
+            internal const ulong July2019Balance = 97950;
+            internal const ulong October2019Balance = 99526;
+            internal const ulong February2020Balance = 102321;
+            internal const ulong May2021Balance = 115190;
+            internal const ulong May2021BalanceHotFix = 115728;
+            internal const ulong June2021Balance = 116210;
+            internal const ulong EODBeta1 = 118697;
+            internal const ulong EODBeta2 = 119939;
+            internal const ulong EODBeta3 = 121168;
+            internal const ulong EODBeta4 = 122479;
+            internal const ulong March2022Balance = 126520;
+            internal const ulong March2022Balance2 = 127285;
+            internal const ulong June2022Balance = 130910;
+            internal const ulong June2022BalanceHotFix = 131084;
+            internal const ulong August2022BalanceHotFix = 132359;
+            internal const ulong August2022Balance = 133322;
+            internal const ulong October2022Balance = 135242;
+            internal const ulong November2022Balance = 137943;
+            //
+            internal const ulong EndOfLife = ulong.MaxValue;
+        }
+
+        internal static class ArcDPSBuilds
+        {
+            internal const int StartOfLife = int.MinValue;
+            //
+            internal const int ProperConfusionDamageSimulation = 20210529;
+            internal const int ScoringSystemChange = 20210800; // was somewhere around there
+            internal const int DirectX11Update = 20210923;
+            internal const int InternalSkillIDsChange = 20220304;
+            internal const int BuffAttrFlatIncRemoved = 20220308;
+            internal const int FunctionalIDToGUIDEvents = 20220709;
+            //
+            internal const int EndOfLife = int.MaxValue;
+        }
+
+        public static class WeaponSetIDs
+        {
+            public const int FirstLandSet = 4;
+            public const int SecondLandSet = 5;
+            public const int FirstWaterSet = 0;
+            public const int SecondWaterSet = 1;
+            public const int TransformSet = 3;
+            public const int KitSet = 2;
+        }
+
         // Activation
         public enum Activation : byte
         {
@@ -277,7 +333,7 @@ namespace GW2EIEvtcParser
         internal static BuffAttribute GetBuffAttribute(short bt, int evtcVersion)
         {
             BuffAttribute res;
-            if (evtcVersion >= ParserHelper.ArcDPSBuilds.BuffAttrFlatIncRemoved)
+            if (evtcVersion >= ArcDPSBuilds.BuffAttrFlatIncRemoved)
             {
                 // Enum has shifted by -1
                 if (bt <= (byte)BuffAttribute.SiphonRec - 1)
@@ -1056,16 +1112,6 @@ namespace GW2EIEvtcParser
         public static MinionID GetMinionID(int id)
         {
             return Enum.IsDefined(typeof(MinionID), id) ? (MinionID)id : MinionID.Unknown;
-        }
-
-        public static bool IsKnownMinionID(AgentItem minion, ParserHelper.Spec spec)
-        {
-            return ProfHelper.IsKnownMinionID(minion, spec);
-        }
-
-        public static bool IsKnownMinionID(AbstractSingleActor minion, ParserHelper.Spec spec)
-        {
-            return IsKnownMinionID(minion.AgentItem, spec);
         }
 
     }
